+++ src/Zarr.jl
@@ -4,7 +4,7 @@ import JSON
 import Blosc
 
 include("metadata.jl")
-include("Compressors.jl")
+include("Compressors/Compressors.jl")
 include("Storage/Storage.jl")
 include("Filters.jl")
 include("ZArray.jl")
